refactor(form): move AttributeList
ref #276
